Move legacy identifier interpreting to separate file from parkfile
