+++ src/openrct2/FileClassifier.cpp
@@ -9,11 +9,11 @@
 
 #include "FileClassifier.h"
 
-#include "ParkFile.h"
 #include "core/Console.hpp"
 #include "core/FileStream.h"
 #include "core/Path.hpp"
 #include "core/String.hpp"
+#include "park/ParkFile.h"
 #include "rct12/SawyerChunkReader.h"
 #include "scenario/Scenario.h"
 #include "util/SawyerCoding.h"
