+++ src/openrct2/ReplayManager.cpp
@@ -13,7 +13,6 @@
 #include "Game.h"
 #include "GameStateSnapshots.h"
 #include "OpenRCT2.h"
-#include "ParkFile.h"
 #include "ParkImporter.h"
 #include "PlatformEnvironment.h"
 #include "actions/FootpathPlaceAction.h"
@@ -31,6 +30,7 @@
 #include "management/NewsItem.h"
 #include "object/ObjectManager.h"
 #include "object/ObjectRepository.h"
+#include "park/ParkFile.h"
 #include "scenario/Scenario.h"
 #include "world/Park.h"
 #include "zlib.h"
