+++ src/openrct2/cmdline/ConvertCommand.cpp
@@ -10,13 +10,13 @@
 #include "../Context.h"
 #include "../FileClassifier.h"
 #include "../OpenRCT2.h"
-#include "../ParkFile.h"
 #include "../ParkImporter.h"
 #include "../common.h"
 #include "../core/Console.hpp"
 #include "../core/Path.hpp"
 #include "../interface/Window.h"
 #include "../object/ObjectManager.h"
+#include "../park/ParkFile.h"
 #include "../scenario/Scenario.h"
 #include "CommandLine.hpp"
