+++ src/openrct2/libopenrct2.vcxproj
@@ -304,8 +304,10 @@
     <ClInclude Include="paint\tile_element\Paint.Surface.h" />
     <ClInclude Include="paint\tile_element\Paint.TileElement.h" />
     <ClInclude Include="paint\VirtualFloor.h" />
-    <ClInclude Include="ParkFile.h" />
     <ClInclude Include="ParkImporter.h" />
+    <ClInclude Include="park\Legacy.h" />
+    <ClInclude Include="park\ParkFile.h" />
+    <ClInclude Include="peep\Guest.h" />
     <ClInclude Include="peep\GuestPathfinding.h" />
     <ClInclude Include="peep\RideUseSystem.h" />
     <ClInclude Include="PlatformEnvironment.h" />
@@ -771,8 +773,9 @@
     <ClCompile Include="paint\tile_element\Paint.TileElement.cpp" />
     <ClCompile Include="paint\tile_element\Paint.Wall.cpp" />
     <ClCompile Include="paint\VirtualFloor.cpp" />
-    <ClCompile Include="ParkFile.cpp" />
     <ClCompile Include="ParkImporter.cpp" />
+    <ClCompile Include="park\Legacy.cpp" />
+    <ClCompile Include="park\ParkFile.cpp" />
     <ClCompile Include="peep\GuestPathfinding.cpp" />
     <ClCompile Include="peep\PeepData.cpp" />
     <ClCompile Include="peep\RideUseSystem.cpp" />
