+++ src/openrct2/network/NetworkBase.cpp
@@ -13,7 +13,6 @@
 #include "../Game.h"
 #include "../GameStateSnapshots.h"
 #include "../OpenRCT2.h"
-#include "../ParkFile.h"
 #include "../PlatformEnvironment.h"
 #include "../actions/LoadOrQuitAction.h"
 #include "../actions/NetworkModifyGroupAction.h"
@@ -24,6 +23,7 @@
 #include "../entity/EntityRegistry.h"
 #include "../entity/EntityTweener.h"
 #include "../localisation/Formatting.h"
+#include "../park/ParkFile.h"
 #include "../platform/Platform2.h"
 #include "../scenario/Scenario.h"
 #include "../scripting/ScriptEngine.h"
