+++ src/openrct2/park/ParkFile.h
