+++ src/openrct2/platform/Crash.cpp
@@ -27,7 +27,6 @@
 #    include "../Context.h"
 #    include "../Game.h"
 #    include "../OpenRCT2.h"
-#    include "../ParkFile.h"
 #    include "../PlatformEnvironment.h"
 #    include "../Version.h"
 #    include "../config/Config.h"
@@ -38,6 +37,7 @@
 #    include "../interface/Screenshot.h"
 #    include "../localisation/Language.h"
 #    include "../object/ObjectManager.h"
+#    include "../park/ParkFile.h"
 #    include "../scenario/Scenario.h"
 #    include "../util/SawyerCoding.h"
 #    include "../util/Util.h"
